Switch to IPFS (#5)
* Setup Fleek storage

* Fix saving bids

* Fix getting the winning bid

* Disallow outbidding with the same amount as the current top bidder
...and some optimizations.

* Use timestamps for tracking the order of bids, optimize getting the winning one

* Refactor ipfs-interactions to a separate file

* Remove valid cycle check from deleting orders
It wasn't implemented anyways and it's probably not even needed. Old bids are periodically deleted on the Solana version too. Here, while that's not needed, disallowing it doesn't bring advantages either.
Since not having this saves us a contract call and decoding order parameters, this could count as an optimization too lol

Co-authored-by: JozsaErik <jozsaerik97@gmail.com>